layers: Move temporary file path code to utils
