+++ layers/core_checks/cc_device.cpp
@@ -20,7 +20,6 @@
  */
 
 #include <fstream>
-#include <sys/stat.h>
 #include <vector>
 
 #if defined(__linux__) || defined(__FreeBSD__) || defined(__OpenBSD__)
@@ -433,20 +432,7 @@ void CoreChecks::CreateDevice(const VkDeviceCreateInfo *pCreateInfo) {
 
     // Allocate shader validation cache
     if (!disabled[shader_validation_caching] && !disabled[shader_validation] && !core_validation_cache) {
-        auto tmp_path = GetEnvironment("XDG_CACHE_HOME");
-        if (!tmp_path.size()) {
-            auto cachepath = GetEnvironment("HOME") + "/.cache";
-            struct stat info;
-            if (stat(cachepath.c_str(), &info) == 0) {
-                if ((info.st_mode & S_IFMT) == S_IFDIR) {
-                    tmp_path = cachepath;
-                }
-            }
-        }
-        if (!tmp_path.size()) tmp_path = GetEnvironment("TMPDIR");
-        if (!tmp_path.size()) tmp_path = GetEnvironment("TMP");
-        if (!tmp_path.size()) tmp_path = GetEnvironment("TEMP");
-        if (!tmp_path.size()) tmp_path = "/tmp";
+        auto tmp_path = GetTempFilePath();
         validation_cache_path = tmp_path + "/shader_validation_cache";
 #if defined(__linux__) || defined(__FreeBSD__) || defined(__OpenBSD__)
         validation_cache_path += "-" + std::to_string(getuid());
